Issue #689: Make IslandLocatorLogic thread-safe (copy location object)

+++ uSkyBlock-Core/src/main/java/us/talabrek/ultimateskyblock/util/TimeUtil.java
@@ -82,7 +82,7 @@ public enum TimeUtil {
         return (int) Math.ceil(millis / 1000f);
     }
 
-    public static long millisAsTicks(int ms) {
+    public static long millisAsTicks(long ms) {
         return ms / 50;
     }
 }
